Refactor the scopes
