+++ src/main/scala/scalajs_java/compiler/Compiler.scala
@@ -8,8 +8,9 @@ import org.scalajs.core.ir.{Position, Trees => irt, Types => irtpe}
 
 import scalajs_java.compiler.passes.ConstructorPass.ConstructorsT
 import scalajs_java.trees._
-import scalajs_java.utils.Scope.ClassMapT
 import scalajs_java.utils._
+import scalajs_java.utils.scope.Scope.ClassMapT
+import scalajs_java.utils.scope._
 
 
 /** Main compiler.
