+++ src/main/scala/scalajs_java/compiler/passes/CompilerPass.scala
@@ -1,13 +1,13 @@
 package scalajs_java.compiler.passes
 
-import org.scalajs.core.ir.{Trees => irt}
 import org.scalajs.core.ir.Trees.ClassDef
+import org.scalajs.core.ir.{Trees => irt}
 
 import scalajs_java.compiler.Compiler
 import scalajs_java.compiler.passes.ConstructorPass.ConstructorsT
-import scalajs_java.trees.{ClassDecl, CompilationUnit, Expr}
-import scalajs_java.utils.{CompilerPhase, ErrorHandler, ScopeElem}
-import scalajs_java.utils.Scope.ClassMapT
+import scalajs_java.trees.{CompilationUnit, Expr}
+import scalajs_java.utils.scope.Scope.ClassMapT
+import scalajs_java.utils.{CompilerPhase, ErrorHandler}
 
 class CompilerPass(inits: Map[String, Expr],
                    classes: ClassMapT,
