+++ src/main/scala/scalajs_java/compiler/passes/ExpSymsPass.scala
@@ -2,7 +2,7 @@ package scalajs_java.compiler.passes
 import scala.collection.mutable.{Map => MMap}
 import scalajs_java.traversals.ExportedSymbolsTraverse
 import scalajs_java.trees.CompilationUnit
-import scalajs_java.utils.Scope.ScopeT
+import scalajs_java.utils.scope.ScopeState
 import scalajs_java.utils.{CompilerPhase, ErrorHandler}
 
 
@@ -16,13 +16,13 @@ class ExpSymsPass(verb: Boolean=false)
   override val errorHandler: ErrorHandler =
     new ErrorHandler(CompilerPhase(name))
 
-  var scope: ScopeT = ScopeT.empty
+  var scope: ScopeState = ScopeState.empty
 
   val expSymTraverse = new ExportedSymbolsTraverse(errorHandler)
 
   override def runPass(tree: CompilationUnit): CompilationUnit = {
     val res = expSymTraverse.traverse(tree)
-    scope = expSymTraverse.scope
+    scope = expSymTraverse.scopeState
     res
   }
 }
