+++ src/main/scala/scalajs_java/compiler/passes/RefTagPass.scala
@@ -2,11 +2,11 @@ package scalajs_java.compiler.passes
 
 import scalajs_java.traversals.RefTraverse
 import scalajs_java.trees.CompilationUnit
-import scalajs_java.utils.Scope.ScopeT
+import scalajs_java.utils.scope.ScopeState
 import scalajs_java.utils.{CompilerPhase, ErrorHandler}
 
 class RefTagPass(override val verbose: Boolean=false,
-                 scope: ScopeT) extends Pass[CompilationUnit, CompilationUnit] {
+                 scope: ScopeState) extends Pass[CompilationUnit, CompilationUnit] {
 
   override val name = "Variable reference tagging"
