+++ src/main/scala/scalajs_java/traversals/RefTraverse.scala
@@ -1,16 +1,16 @@
 package scalajs_java.traversals
 
 import scalajs_java.trees._
-import scalajs_java.utils.Scope.ScopeT
-import scalajs_java.utils.{ErrorHandler, MethodElem, Scope}
+import scalajs_java.utils.scope.{MethodElem, Scope, ScopeState}
+import scalajs_java.utils.ErrorHandler
 
 /** Traverses the tree constructing the scope and tagging
   * `Ident` nodes with the tree nodes they reference. */
-class RefTraverse(errHandler: ErrorHandler, initScope: ScopeT) extends Traverse with Scope { self =>
+class RefTraverse(errHandler: ErrorHandler, initScope: ScopeState) extends Traverse with Scope { self =>
 
   override val errorHanlder: ErrorHandler = errHandler
 
-  scope = initScope
+  scopeState = initScope
 
   override def traverse(classDecl: ClassDecl): ClassDecl = {
     withScope[ClassDecl, ClassDecl](classDecl.members, classDecl)(super.traverse)
@@ -35,7 +35,7 @@ class RefTraverse(errHandler: ErrorHandler, initScope: ScopeT) extends Traverse 
 
   override def traverse(ident: Ident): Ident = {
     val symbol = ident.name.str
-    val referredTree = scope.getElem(symbol)
+    val referredTree = scopeState.getElem(symbol)
 
     ident.copy(refVar = referredTree)(ident.pos)
   }
@@ -45,13 +45,13 @@ class RefTraverse(errHandler: ErrorHandler, initScope: ScopeT) extends Traverse 
 
     val refTree: Option[MethodElem] = methodInv.methodSel match {
       case FieldAccess(name, _, _, _) =>
-        scope.getMethod(name)
+        scopeState.getMethod(name)
 
       case Ident(_, name, _, rv, _) =>
         if (rv.isDefined) rv.flatMap {
           case mElem: MethodElem => Some(mElem)
           case _                 => None
-        } else scope.getMethod(name)
+        } else scopeState.getMethod(name)
 
       case _ =>
         None
