--- src/main/scala/scalajs_java/utils/Scope.scala
@@ -1,191 +1,0 @@
-package scalajs_java.utils
-
-import scala.collection.mutable.{Map => MMap}
-import scalajs_java.trees._
-import org.scalajs.core.ir.{Trees => irt}
-
-import scalajs_java.compiler.Utils
-
-sealed trait ScopeElem {
-  val name: String
-  val decl: Tree
-  val kind: VarKind
-}
-
-sealed trait MethodElem extends ScopeElem {
-  override val kind: VarKind = Method
-}
-
-case class VarInfo(name: String, mangled: irt.Ident, decl: VarDecl, kind: VarKind) extends ScopeElem
-
-case class MethodInfo(name: String, decl: MethodDecl) extends MethodElem
-
-case class ClassInfo(name: String, decl: ClassDecl, kind: VarKind=Class) extends ScopeElem
-
-case class LibraryMethod(name: String) extends MethodElem {
-  override val decl: Tree = Skip()(Position.noPosition)
-}
-
-trait Scope {
-
-  import scalajs_java.utils.Scope.ScopeT
-
-  val errorHanlder: ErrorHandler
-
-  var scope: ScopeT = ScopeT.empty
-
-  val mangler = new Mangler
-
-  def addToScope(scopeElem: ScopeElem): Unit = scope.addElem(scopeElem)
-
-  def remFromScope(symbol: String): Unit = scope.remElem(symbol)
-
-  def getScopeElems(members: List[Tree]): List[ScopeElem] = {
-    members.collect {
-      case vd: VarDecl =>
-        val pos = Utils.getPosition(vd)
-        val mangledName = vd.kind match {
-          case LocalVar    => mangler.encodeLocalSym(vd.symbol)(pos)
-          case Param       => mangler.encodeParamIdent(vd.symbol)(pos)
-          case ClassMember => mangler.encodeFieldSym(vd.symbol)(pos)
-          case _           => irt.Ident(vd.symbol.toString)(pos)
-        }
-        VarInfo(vd.name.str, mangledName, vd, vd.kind)
-
-      case md: MethodDecl =>
-        MethodInfo(md.name.str, md)
-    }
-  }
-
-  /** The key method of this module, used to perform
-    * computations with a temporary scope update */
-  def withScope[T, S](scope: List[Tree], tree: T)(f: T => S): S = {
-    val scopeElems = getScopeElems(scope)
-    scopeElems.foreach(addToScope)
-
-    val res = f(tree)
-
-    scopeElems.foreach(vi => remFromScope(vi.name))
-    res
-  }
-
-  def withScope[T, S](scopeElem: Tree, tree: T)(f: T => S): S =
-    withScope[T, S](List(scopeElem), tree)(f)
-}
-
-object Scope {
-
-  class ScopeT(val vars: MMap[String, List[VarInfo]],
-               val methods: MMap[String, List[MethodElem]],
-               val classes: MMap[String, List[ClassInfo]]) {
-
-    def ++(that: ScopeT): ScopeT = {
-      new ScopeT(
-        this.vars ++ that.vars,
-        this.methods ++ that.methods,
-        this.classes ++ that.classes)
-    }
-
-   def addElem(scopeElem: ScopeElem): Unit = {
-     val sym = scopeElem.name
-     scopeElem match {
-       case vi: VarInfo =>
-         if (!vars.contains(sym)) vars(sym) = Nil
-
-         vars(sym) = vi :: vars(sym)
-
-       case mi: MethodInfo =>
-         if (!methods.contains(sym)) methods(sym) = Nil
-
-         methods(sym) = mi :: methods(sym)
-
-       case ci: ClassInfo =>
-         if (!classes.contains(sym)) classes(sym) = Nil
-
-         classes(sym) = ci :: classes(sym)
-
-       case _: LibraryMethod =>
-         throw new Exception("Cannot compile library methods")
-     }
-   }
-
-    def remElem(sym: String): Unit = {
-      if (vars.contains(sym)) {
-        if (vars(sym).nonEmpty)
-          vars(sym) = vars(sym).tail
-
-        if (vars(sym).isEmpty)
-          vars.remove(sym)
-      } else if (methods.contains(sym)) {
-        if (methods(sym).nonEmpty)
-          methods(sym) = methods(sym).tail
-
-        if (methods(sym).isEmpty)
-          methods.remove(sym)
-      } else if (classes.contains(sym)) {
-        if (classes(sym).nonEmpty)
-          classes(sym) = classes(sym).tail
-
-        if (classes(sym).isEmpty)
-          classes.remove(sym)
-      } else {
-        throw new Exception(s"[ScopeT -- remElem]" +
-            s"Cannot remove element ($sym) -- not in scope.")
-      }
-    }
-
-    def getMethod(sym: String): Option[MethodElem] = {
-      methods.get(sym).flatMap {
-        case head :: _ => Some(head)
-        case _         => Scope.libraryMethods.get(sym)
-      }
-    }
-
-    def getVar(sym: String): Option[VarInfo] = {
-      vars.get(sym).flatMap {
-        case head :: _ => Some(head)
-        case _         => None
-      }
-    }
-
-    def getClass(sym: String): Option[ClassInfo] = {
-      classes.get(sym).flatMap {
-        case head :: _ => Some(head)
-        case _         => None
-      }
-    }
-
-    def getElem(sym: String): Option[ScopeElem] = {
-      if (vars.contains(sym)) getVar(sym)
-      else if (methods.contains(sym)) getMethod(sym)
-      else if (classes.contains(sym)) getClass(sym)
-      else None
-    }
-  }
-
-  object ScopeT {
-    def empty: ScopeT =
-      new ScopeT(MMap.empty, MMap.empty, MMap.empty)
-  }
-
-  type ClassMapT = Map[String, ClassDecl]
-
-  def empty: ScopeT = ScopeT.empty
-
-  def mkScope(scopes: List[ScopeT]): ScopeT =
-    scopes.reduce {_ ++ _}
-
-  def getClasses(scope: ScopeT): ClassMapT = {
-    scope.classes
-        .filter(entry => entry._2.nonEmpty)
-        .map(entry => (entry._1, entry._2.head.decl))
-        .toMap
-  }
-
-  /* TODO make only `System.out.println` a library method,
-   * not every `println` */
-  val libraryMethods = Map[String, MethodElem](
-    "println" -> LibraryMethod("println"),
-    "print" -> LibraryMethod("print")
-  )
-}
